Publish ssp v2 calculator
